// src/Screens/AddMedicalInfo/types.ts
export type AddMedicalInfoFormType = {
  medicalInfoType: string;

  medicationName?: string;
  dosage?: string;
  frequency?: string;
  notesMedication?: string;

  conditionName?: string;
  diagnosisDate?: string;
  notesCondition?: string;

  procedureName?: string;
  surgeryDate?: string;
  surgeon?: string;
  reason?: string;

  allergy?: string;
  reaction?: string;
  severity?: string;
  notesAllergy?: string;

  
};